requirements.txt
numpy==1.26.4
scipy
tensorboard
librosa==0.9.2
numba==0.59.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
pytorch-lightning==2.2.1
gradio>=5.0
onnxruntime; sys_platform == 'darwin'
onnxruntime-gpu; sys_platform != 'darwin'
tqdm
pyarmor==8.5.11
funasr==1.0.27
cn2an
pypinyin
pyopenjtalk>=0.3.4
g2p_en
torchaudio==2.0.2
modelscope==1.10.0
sentencepiece
transformers
chardet
PyYAML
psutil
jieba_fast
jieba
LangSegment>=0.2.0
Faster_Whisper
wordsegment
rotary_embedding_torch
pyjyutping 
g2pk2
ko_pron
opencc; sys_platform != 'linux'
opencc==1.1.1; sys_platform == 'linux'
python_mecab_ko; sys_platform != 'win32'

ctranslate2==4.4.0
ffmpeg==1.4
ffmpeg-python==0.2.0

ultralytics
opencv-python
Pillow
qrcode
netifaces
yacs
einops
scikit-image

tb-nightly -i https://mirrors.aliyun.com/pypi/simple
facexlib
gfpgan
